Split driver methods into mixin classes
